feat: update precision

+++ train_k_gpt.py
@@ -54,7 +54,7 @@ if __name__ == '__main__':
         trainer = pl.Trainer(accelerator=args.accelerator, devices=args.devices, num_nodes=args.num_nodes,
                              strategy=DDPStrategy(find_unused_parameters=False, gradient_as_bucket_view=True),
                              callbacks=[model_checkpoint, lr_monitor], max_epochs=args.max_epochs,
-                             precision="bf16-mixed", accumulate_grad_batches=args.grad_batch_size)
+                             precision="32-true", accumulate_grad_batches=args.grad_batch_size)
         trainer.fit(model, datamodule, ckpt_path=args.ckpt_path)
     else:
         model = KGPT.load_from_checkpoint(
